server/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
Pillow>=10.0.0
inky>=0.0.3
